Refactored config and readers
